+++ src/interfaces/ChartData.interface.ts
@@ -8,7 +8,7 @@ export interface ChartSeries {
   /** key - the name of the series */
   key: string;
   /** values - represents the array of {x, y} data points */
-  values: Array<ChartPoint>;
+  values: ChartPoint[];
   /** color - optional: choose your own line color (e.g.'#ff7f0e') */
   color?: string;
   /** area - optional: set to true if you want this line to turn into a filled area chart */
@@ -17,4 +17,7 @@ export interface ChartSeries {
   shape?: string;
   /** bar - optional: set true in LinePLusBarChart for select bar option */
   bar?: boolean;
+
+  strokeWidth?: any;
+  classed?: any;
 }